Consolodated dashbord components
